added docker to repo and ui-docker.yml

+++ src/config/global.setup.ts
@@ -2,8 +2,7 @@ import path from "path";
 import { rimraf } from "rimraf";
 
 export default function globalSetup() {
+  if (process.env.CI) return;
   const resultsPath = path.resolve(__dirname, "allure-results");
   rimraf.sync(resultsPath);
-
-  console.log("from global setup");
 }
